Tests: polyfill select PHPUnit 9.1 functions
Use a trait to selectively polyfill a number of methods introduced in PHPUnit 9.1.

These new PHPUnit methods all replace already existing methods, effectively renaming the old methods.

The trait only contains polyfill methods for assertion methods currently used in the WordPress test suite, but can be expanded to cover the complete set of PHPUnit 9.1 renamed methods.

The polyfill methods themselves are effectively just a placeholder/wrapper, they will use the PHPUnit native functions to execute the assertions.

With the polyfill in place, the WP test suite can already be upgraded to use the _current_ name of these functions, rather than the old name.

Once the need to support PHPUnit < 9.1 is dropped, the polyfill can be removed and that is all that is needed at that time.

Ref:
* https://github.com/sebastianbergmann/phpunit/blob/9.1.5/ChangeLog-9.1.md#910---2020-04-03