WIP - needs extra methods, tests + fix commit msg | WP_Comment: fix existing magic methods + add tests
I've reviewed the pre-existing magic methods in the `WP_Comment` class with an eye on the PHP 8.2 dynamic properties deprecation.

The class only implemented the `__isset()` and `__get()` method which created an unstable interface.

/*
The methods were _mostly_ correct, but not completely so:
* The `__get()` method would throw an "Undefined property" notice if one of the "compat" allow listed properties without a default value would be requested prior to a value being assigned.
    This has been fixed by adding an `isset()` check.
* The `__get()` method also had an inconsistent return as it would potentially return a value, but for undeclared properties and properties not on the "allow list", it would be a `void` method.
    This was incorrect as a magic `__get()` method should always return a value, even when that value is `null`.
    Fixed now by adding the fall-back `return null`.
* The magic `__set()` method MUST be a `void` method, but was returning the new value.
    Fixed now by removing the `return`.
*/

Note: while the magic method requirements were not that strict in older PHP versions, as of PHP 8.0, they have become stricter, so magic methods should really comply with the PHP requirements for the methods.

As this class appears to not have any tests yet, I have set up a new test class with an initial set of tests covering the magic methods.

Refs:
* https://www.php.net/manual/en/language.oop5.overloading.php#language.oop5.overloading.members
* https://wiki.php.net/rfc/magic-methods-signature