wpdb: add additional tests for the magic methods
Note: these tests use a partial mock of the `wpdb` class to prevent these tests changing the global `$wpdb` object, which could affect the results of other tests.

What I mean by a partial mock is that:
* The constructor is being skipped when the mock is initialized.
    This is done for two reasons:
    1. The constructor sets the value of quite a few properties, including security sensitive ones like `$dbpassword`. We don't need these properties to be set to "real" values for these tests to do their work.
    2. The constructor calls quite a few methods in the class, which are unrelated/unnecessary for these tests, including `db_connect()`.
        This means that a) all those methods would need to be mocked and b) the tests would be unnecessarily slow due to making a new db connection for every single test.
        Skipping the constructor prevents all that.
* One method which is actually used in the methods under test (`load_col_info()`) is being mocked. All other methods use the original code in `wpdb`.

The mocking does have one implication: the class used to test these methods is now a _child class_ to `wpdb` and not the `wpdb` parent itself.