[DISCUSS] wpdb::__unset(): bug fix
There were three properties, which were introduced in [30345](https://core.trac.wordpress.org/changeset/30345), which were protected from being overwritten (set).

That protection, however, was not extended to the `__unset()` method, leaving these properties only partially protected.
IMO this is an oversight (bug) in the initial implementation.

This commit fixes this by applying the same protection as was already given to these properties for `__set()`, to these properties in the `__unset()` method.

Includes updating the unit tests to match.

Note: this _could_ be considered a BC-break, in which case, an alternative implementation would be to throw a `_doing_it_wrong()` and to continue to unset the property when so requested.