Tests_Admin_WpListTable: remove incorrect `@covers` tag
This class does not contain tests to covers _all_ methods and functionality in `WP_list_Table`, so the test class should not be marked to cover the whole class.

The one test which is included in this test class already has a more specific `@covers` tag, so this class level one should be removed.