Tests_Admin_WpListTable: minor tweak to make adding additional tests easier
`WP_Screen` needs a hook suffix and is set up in the `__construct()` of the `WP_List_Table` class, so all tests which test `WP_List_Table` will need for the hook suffix to be set.

By moving this test to a `set_up()` method, all (current and future) tests will benefit from it.