Build/Test Tools: Enable running the tests on PHP 8.4
:warning: Please read Trac ticket 62061 - this PR can't be committed until other patches have been committed first! :warning:

PHP 8.4 is expected to be released at the end of November 2024.

Enabling the tests to run in CI on PHP 8.4 allows WordPress core to start getting ready.

Note: Xdebug-related tests will not be run on PHP 8.4 at this time as the Docker image for PHP 8.4 does not contain Xdebug yet. Once a stable release of Xdebug 3.4.0 is available, it can be added to the Docker image and the test step can then be enabled for PHP 8.4.